refactor judge to reduce hacky imports
